requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==0.25.3
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
Werkzeug==0.16.0
